not expecting more than 2^32 datasets in a given folder

+++ core/src/main/java/org/srs/datacat/shared/container/BasicStat.java
@@ -10,6 +10,7 @@ import javax.xml.bind.annotation.XmlSeeAlso;
 import javax.xml.bind.annotation.XmlType;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
+
 /**
  * Basic stat info for a group or a folder
  * 
@@ -29,13 +30,13 @@ public class BasicStat {
         DATASET;
     }
 
-    private long datasetCount;
-    private long groupCount;
-    private long folderCount;
+    private int datasetCount;
+    private int groupCount;
+    private int folderCount;
     
     public BasicStat(){}
     
-    public BasicStat(long folderCount, long groupCount, long datasetCount){
+    public BasicStat(int folderCount, int groupCount, int datasetCount){
         this.datasetCount = datasetCount;
         this.groupCount = groupCount;
         this.folderCount = folderCount;
@@ -46,39 +47,39 @@ public class BasicStat {
     }
     
     @XmlElement(required=true)
-    public Long getChildCount(){
+    public Integer getChildCount(){
         return folderCount + groupCount + datasetCount;
     }
     
     @XmlElement(required=false)
-    public Long getChildContainerCount(){
+    public Integer getChildContainerCount(){
         return folderCount + groupCount;
     }
     
     @XmlElement(required=true)
-    public Long getDatasetCount(){
+    public Integer getDatasetCount(){
         return datasetCount;
     }
     
     @XmlElement(required=true)
-    public Long getGroupCount(){
+    public Integer getGroupCount(){
         return groupCount;
     }
     
     @XmlElement(required=true)
-    public Long getFolderCount(){
+    public Integer getFolderCount(){
         return folderCount;
     }
     
-    public void setDatasetCount(long datasets){
+    public void setDatasetCount(int datasets){
         this.datasetCount = datasets;
     }
     
-    public void setGroupCount(long groups){
+    public void setGroupCount(int groups){
         this.groupCount = groups;
     }
     
-    public void setFolderCount(long folders){
+    public void setFolderCount(int folders){
         this.folderCount = folders;
     }
